requirements.txt
# 核心依赖
requests>=2.25.0
Pillow
pypdfium2>=4.0.0
openpyxl>=3.0.0
anthropic>=0.75.0
psutil>=5.8.0

# GUI相关
tkinter  # 通常随Python安装

# 打包工具
PyInstaller

# 开发和测试工具（可选）
pytest